Name argument solves my attribute problem
Although it doesn't solve my attribute problem in general, in this case
I don't need to create my own one, because I get one from the system for
free.

Next problem is that my ray starts from the center of the player, so I
can turn it 90* around without having to figure out all the math. The
disadvantage is that if I stand on an area with collision detection,
then it also detects a collision when I'm inside the area and doesn't
let me move out in any direction. So I either need to deactivate the
info area, or figure out a better way to raycast.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>