The info pops up correctly
... most of the time. And pushing the box works, most of the time.
Now we get to the fun part of debugging.

I'm personally a little disappointed that I have to edit the instance
specific button handling in if-else blocks. Maybe I will find a way to
add this functionality to a menu later in some way. E.g. the buttons
raise signals and the menu responds based on the signal received?

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>